MCP-1554: move services to proper packages
