+++ app/src/test/java/gov/va/vro/VroApplicationTest.java
@@ -1,0 +1,11 @@
+package gov.va.vro;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class VroApplicationTest {
+
+  @Test
+  void contextLoads() {}
+}
