+++ service/save-to-db/src/main/java/gov/va/vro/service/db/SaveToDbServiceImpl.java
@@ -1,13 +1,13 @@
 package gov.va.vro.service.db;
 
-import gov.va.starter.example.service.spi.db.SaveToDbService;
-import gov.va.starter.example.service.spi.db.model.Claim;
 import gov.va.vro.persistence.model.ClaimEntity;
 import gov.va.vro.persistence.model.ContentionEntity;
 import gov.va.vro.persistence.model.VeteranEntity;
 import gov.va.vro.persistence.repository.ClaimRepository;
 import gov.va.vro.persistence.repository.VeteranRepository;
 import gov.va.vro.service.db.mapper.ClaimMapper;
+import gov.va.vro.service.spi.db.SaveToDbService;
+import gov.va.vro.service.spi.db.model.Claim;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
