+++ service/save-to-db/src/main/java/gov/va/vro/service/db/mapper/ClaimMapper.java
@@ -1,7 +1,7 @@
 package gov.va.vro.service.db.mapper;
 
-import gov.va.starter.example.service.spi.db.model.Claim;
 import gov.va.vro.persistence.model.ClaimEntity;
+import gov.va.vro.service.spi.db.model.Claim;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
