--- service/spi/src/main/java/gov/va/starter/example/service/spi/db/SaveToDbService.java
@@ -1,7 +1,0 @@
-package gov.va.starter.example.service.spi.db;
-
-import gov.va.starter.example.service.spi.db.model.Claim;
-
-public interface SaveToDbService {
-  Claim insertClaim(Claim claim);
-}
