+++ service/spi/src/main/java/gov/va/vro/service/spi/db/SaveToDbService.java
@@ -1,0 +1,7 @@
+package gov.va.vro.service.spi.db;
+
+import gov.va.vro.service.spi.db.model.Claim;
+
+public interface SaveToDbService {
+  Claim insertClaim(Claim claim);
+}
