+++ service/spi/src/main/java/gov/va/vro/service/spi/db/model/Claim.java
@@ -1,4 +1,4 @@
-package gov.va.starter.example.service.spi.db.model;
+package gov.va.vro.service.spi.db.model;
 
 import lombok.Data;
